+++ Engine/MovingFloor.h
@@ -8,15 +8,15 @@ class MovingFloor
 {
 public:
 	MovingFloor( const Vec2& pos,float angle,
-		float leftStop,float rightStop,float speed );
+		int leftStop,int rightStop,float speed );
 
 	void Update( float dt ) override;
 	void Draw( Graphics& gfx ) const override;
 private:
 	void MoveBy( const Vec2& amount );
 private:
-	float leftPos;
-	float rightPos;
+	int leftPos;
+	int rightPos;
 	Vec2 vel;
 	int dir = -1;
 };
